updates landing page with calender
